+++ cookbooks/arcgis-enterprise/attributes/webadaptor.rb
@@ -30,9 +30,14 @@ default['arcgis']['web_adaptor'].tap do |web_adaptor|
     web_adaptor['install_dir'] = ''
 
     case node['arcgis']['version']
+    when '10.6'
+      web_adaptor['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
+                                            'Web_Adaptor_for_Microsoft_IIS_106_161833.exe')
+      web_adaptor['product_code'] = '{4FB9D475-9A23-478D-B9F7-05EBA2073FC7}'
+      web_adaptor['product_code2'] = '{38DBD944-7F0E-48EB-9DCB-98A0567FB062}'
     when '10.5.1'
       web_adaptor['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
-                                            'Web_Adaptor_for_Microsoft_IIS_1051_XXXXXX.exe')
+                                            'Web_Adaptor_for_Microsoft_IIS_1051_156367.exe')
       web_adaptor['product_code'] = '{0A9DA130-E764-485F-8C1A-AD78B04AA7A4}'
       web_adaptor['product_code2'] = '{B8A6A873-ED78-47CE-A9B4-AB3192C47604}'
     when '10.5'
@@ -59,7 +64,11 @@ default['arcgis']['web_adaptor'].tap do |web_adaptor|
                                     '{F53FEE2B-54DD-4A6F-8545-6865F4FBF6DC}',
                                     '{475ACDE5-D140-4F10-9006-C804CA93D2EF}',
                                     '{87B4BD93-A5E5-469E-9224-8A289C6B2F10}',
-                                    '{604CF558-B7E1-4271-8543-75E260080DFA}']
+                                    '{604CF558-B7E1-4271-8543-75E260080DFA}',
+                                    '{0A9DA130-E764-485F-8C1A-AD78B04AA7A4}',
+                                    '{B8A6A873-ED78-47CE-A9B4-AB3192C47604}',
+                                    '{4FB9D475-9A23-478D-B9F7-05EBA2073FC7}',
+                                    '{38DBD944-7F0E-48EB-9DCB-98A0567FB062}']
   else # node['platform'] == 'linux'
     web_adaptor['setup'] = ::File.join(node['arcgis']['repository']['setups'],
                                        node['arcgis']['version'],
@@ -67,9 +76,12 @@ default['arcgis']['web_adaptor'].tap do |web_adaptor|
     web_adaptor['lp-setup'] = node['arcgis']['web_adaptor']['setup']
 
     case node['arcgis']['version']
+    when '10.6'
+      web_adaptor['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
+                                                 'Web_Adaptor_Java_Linux_106_161911.tar.gz')
     when '10.5.1'
       web_adaptor['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
-                                                 'Web_Adaptor_Java_Linux_1051_XXXXXX.tar.gz')
+                                                 'Web_Adaptor_Java_Linux_1051_156442.tar.gz')
     when '10.5'
       web_adaptor['setup_archive'] = ::File.join(node['arcgis']['repository']['archives'],
                                                  'Web_Adaptor_Java_Linux_105_154055.tar.gz')